Rename string and number input and fix comments
